Also extend ClassMethods
This is what will actually happen at runtime, as seen here:

    https://github.com/protocolbuffers/protobuf/blob/7ccf4d8f67878a6ceb2184df279478cb3314372b/ruby/ext/google/protobuf_c/message.c#L1131-L1133

And when we use reflection (e.g., `srb rbi hidden-definitions`) to
discover things that are missing statically, we see many, many lines
that look like this in the generated RBIs. This should cut down on a lot
of that.