Don't attempt to redeclare nesting scope
Sorbet already knows how to handle

    module A::B::C
    end

out of the box. It doesn't require the programmer to write

    module A; end
    module A::B; end
    module A::B::C
    end

like you would have to if you were running this code like normal Ruby.

And this is important, because if it's ever the case that somewhere else
`A` or `B` are actually defined as classes, Sorbet is completely fine
with that:

    # OK

    module A; end
    class A::B; end
    module A::B::C
    end

But if Sorbet ever sees conflicting definitions, then it will complain:

    # ERROR

    module A; end
    class A; end

So we don't gain anything by having these declared explicitly, and also
in some cases it causes problems.