Remove redundant `include`
This `include` is not actually there in message classes at runtime.

The only methods defined on this module are singleton class methods,
which means that even if it were included on these modules somehow, it
wouldn't add any new methods to the class, because singleton class
methods on modules are never accessible to a child module.